refactor(gamelift): split deployer.go into focused domain files
Moves ~300 lines from deployer.go into four focused files while keeping
the public API and behavior identical:

- deployer_fleet.go    — fleet create/wait/delete + GetFleetStatus
- deployer_container.go — container group definition create/wait/delete
- deployer_iam.go      — ensureIAMRole + deleteIAMRole
- deployer_helpers.go  — constants, buildCreateFleetInput, findContainerFleetID

deployer.go now holds only types, constructor, resourceTags, and the
top-level orchestration methods (CreateContainerGroupDefinition,
CreateFleet, Destroy).
