+++ internal/gamelift/deployer_iam.go
@@ -1,0 +1,80 @@
+package gamelift
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/iam"
+	"github.com/devrecon/ludus/internal/awsutil"
+	"github.com/devrecon/ludus/internal/tags"
+)
+
+// ensureIAMRole creates the GameLift fleet IAM role if it doesn't exist, returns the role ARN.
+func (d *Deployer) ensureIAMRole(ctx context.Context) (string, error) {
+	getOut, err := d.iamClient.GetRole(ctx, &iam.GetRoleInput{
+		RoleName: aws.String(iamRoleName),
+	})
+	if err == nil {
+		return aws.ToString(getOut.Role.Arn), nil
+	}
+
+	assumeRolePolicy := `{
+  "Version": "2012-10-17",
+  "Statement": [{
+    "Effect": "Allow",
+    "Principal": {"Service": "gamelift.amazonaws.com"},
+    "Action": "sts:AssumeRole"
+  }]
+}`
+
+	createOut, err := d.iamClient.CreateRole(ctx, &iam.CreateRoleInput{
+		RoleName:                 aws.String(iamRoleName),
+		AssumeRolePolicyDocument: aws.String(assumeRolePolicy),
+		Description:              aws.String("IAM role for Ludus GameLift container fleet"),
+		Tags:                     tags.ToIAMTags(d.resourceTags()),
+	})
+	if err != nil {
+		return "", fmt.Errorf("creating IAM role: %w", err)
+	}
+
+	_, err = d.iamClient.AttachRolePolicy(ctx, &iam.AttachRolePolicyInput{
+		RoleName:  aws.String(iamRoleName),
+		PolicyArn: aws.String(iamPolicyARN),
+	})
+	if err != nil {
+		return "", fmt.Errorf("attaching policy to role: %w", err)
+	}
+
+	// IAM changes take ~10s to propagate globally before GameLift can assume the role.
+	time.Sleep(10 * time.Second)
+
+	return aws.ToString(createOut.Role.Arn), nil
+}
+
+func (d *Deployer) deleteIAMRole(ctx context.Context) error {
+	fmt.Println("Deleting IAM role...")
+
+	_, err := d.iamClient.DetachRolePolicy(ctx, &iam.DetachRolePolicyInput{
+		RoleName:  aws.String(iamRoleName),
+		PolicyArn: aws.String(iamPolicyARN),
+	})
+	if err != nil && !awsutil.IsNotFound(err) {
+		return fmt.Errorf("detaching policy from role: %w", err)
+	}
+
+	_, err = d.iamClient.DeleteRole(ctx, &iam.DeleteRoleInput{
+		RoleName: aws.String(iamRoleName),
+	})
+	if err != nil {
+		if awsutil.IsNotFound(err) {
+			fmt.Println("IAM role not found, skipping.")
+			return nil
+		}
+		return fmt.Errorf("deleting IAM role: %w", err)
+	}
+
+	fmt.Println("IAM role deleted.")
+	return nil
+}
